+++ clients/java/src/main/java/io/camunda/client/impl/command/AgentInstanceHistoryMapper.java
@@ -21,8 +21,8 @@ import io.camunda.client.api.command.AgentInstanceHistoryContent.ObjectContent;
 import io.camunda.client.api.command.AgentInstanceHistoryContent.TextContent;
 import io.camunda.client.api.command.AgentInstanceHistoryMetrics;
 import io.camunda.client.api.command.AgentInstanceHistoryToolCall;
+import io.camunda.client.api.command.AgentInstanceLimits;
 import io.camunda.client.api.command.AgentTool;
-import io.camunda.client.api.command.UpdateAgentInstanceCommandStep1.AgentInstanceLimits;
 import io.camunda.client.api.response.DocumentMetadata;
 import io.camunda.client.api.response.DocumentReferenceResponse;
 import io.camunda.client.protocol.rest.AgentInstanceDocumentContent;
